Making public
moving this project from a private repo to a public repo. I don't feel like scrubbing all the history to double check for security issues back in the early days, so I'm just moving the code as-is to a fresh repo.   (actually the repo I made for this before I decided to start the project in a private repo)